requirements.txt
pyocr
Pillow
